+++ src/app/component-library/conditions-for-use-dialog/conditions-for-use-dialog.module.ts
@@ -1,9 +1,3 @@
-/*
- * Copyright (c) 2021-2024. sovity GmbH
- * Copyright (c) 2024. Fraunhofer Institute for Applied Information Technology FIT
- * Contributors:
- *    - Fraunhofer FIT: Internationalization and German Localization
- */
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {FormsModule} from '@angular/forms';
@@ -11,20 +5,19 @@ import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
 import {MatDialogModule} from '@angular/material/dialog';
 import {MatIconModule} from '@angular/material/icon';
-import {TranslateModule} from '@ngx-translate/core';
 import {NgxJsonViewerModule} from 'ngx-json-viewer';
 import {ConfirmationDialogModule} from '../confirmation-dialog/confirmation-dialog.module';
 import {MarkdownDescriptionModule} from '../markdown-description/markdown-description.module';
 import {PipesAndDirectivesModule} from '../pipes-and-directives/pipes-and-directives.module';
-import {UrlListDialogComponent} from './url-list-dialog/url-list-dialog.component';
-import {UrlListDialogService} from './url-list-dialog/url-list-dialog.service';
+import {ConditionsForUseDialogComponent} from './conditions-for-use-dialog/conditions-for-use-dialog.component';
+import {ConditionsForUseDialogService} from './conditions-for-use-dialog/conditions-for-use-dialog.service';
 
 @NgModule({
   imports: [
     // Angular
     CommonModule,
     FormsModule,
-    TranslateModule,
+
     // Angular Material
     MatButtonModule,
     MatCardModule,
@@ -39,8 +32,8 @@ import {UrlListDialogService} from './url-list-dialog/url-list-dialog.service';
     ConfirmationDialogModule,
     MarkdownDescriptionModule,
   ],
-  declarations: [UrlListDialogComponent],
-  providers: [UrlListDialogService],
-  exports: [UrlListDialogComponent],
+  declarations: [ConditionsForUseDialogComponent],
+  providers: [ConditionsForUseDialogService],
+  exports: [ConditionsForUseDialogComponent],
 })
-export class UrlListDialogModule {}
+export class ConditionsForUseDialogModule {}