+++ src/app/component-library/data-address/data-address-type-select/data-address-type-select-items.ts
@@ -1,12 +1,14 @@
+import {ActiveFeatureSet} from 'src/app/core/config/active-feature-set';
 import {DataAddressTypeSelectItem} from './data-address-type-select-item';
 import {DataAddressTypeSelectMode} from './data-address-type-select-mode';
 
 export const dataAddressTypeSelectItems = (
   type: DataAddressTypeSelectMode,
+  activeFeatureSet: ActiveFeatureSet,
 ): DataAddressTypeSelectItem[] => {
   const items: DataAddressTypeSelectItem[] = [];
 
-  if (type.startsWith('Datasource')) {
+  if (type.startsWith('Datasource') && activeFeatureSet.hasMdsFields()) {
     items.push({
       id: 'On-Request',
       label: '"On Request" Data Offer',