+++ src/app/component-library/policy-editor/model/policy-verbs.ts
@@ -21,26 +21,26 @@ export interface PolicyVerbConfig {
 
 export const SUPPORTED_POLICY_VERBS: PolicyVerbConfig[] = [
   {
-    operandLeftId: policyLeftExpressions.policyEvaluationTime,
-    operandLeftTitle: 'Evaluation Time',
-    operandLeftDescription:
-      'Time at which the policy is evaluated. This can be used to restrict the data offer to certain time periods',
-    supportedOperators: ['GEQ', 'LEQ', 'GT', 'LT'],
-    operandRightType: 'DATE',
-    operandRightPlaceholder: 'MM/DD/YYYY',
-    operandRightHint: 'MM/DD/YYYY',
-    adapter: localDateAdapter,
-  },
-  {
     operandLeftId: policyLeftExpressions.referringConnector,
-    operandLeftTitle: 'Participant ID',
+    operandLeftTitle: "Consumer's Participant ID",
     operandLeftDescription:
-      'Participant ID, also called Connector ID, of the counter-party connector.',
+      "Consumer's Participant ID, also called Connector ID, of the counter-party connector.",
     operandRightType: 'PARTICIPANT_ID',
     supportedOperators: ['EQ', 'IN'],
     operandRightPlaceholder: 'MDSL1234XX.C1234YY',
     operandRightHint: 'Multiple values can be joined by comma',
     adapter: stringArrayOrCommaJoinedAdapter,
+  },
+  {
+    operandLeftId: policyLeftExpressions.policyEvaluationTime,
+    operandLeftTitle: 'Time Restriction',
+    operandLeftDescription:
+      'Time at which the policy is evaluated. This can be used to restrict the data offer to certain time periods',
+    supportedOperators: ['GEQ', 'LEQ', 'GT', 'LT'],
+    operandRightType: 'DATE',
+    operandRightPlaceholder: 'DD/MM/YYYY',
+    operandRightHint: 'DD/MM/YYYY',
+    adapter: localDateAdapter,
   },
 ];